'Refactored by Sourcery'

+++ utils/android/adb_clean.py
@@ -14,7 +14,6 @@ import os
 from adb.commands import DEVICE_TEMP_DIR, reboot, rmdir
 
 
-if __name__ == '__main__':
-    if 'SKIP_ANDROID_CLEAN' not in os.environ:
-        reboot()
-        rmdir(DEVICE_TEMP_DIR)
+if __name__ == '__main__' and 'SKIP_ANDROID_CLEAN' not in os.environ:
+    reboot()
+    rmdir(DEVICE_TEMP_DIR)
